commit homework

+++ homework/homework19/DirDelete.java
@@ -2,13 +2,13 @@ package com.proftelran.homework.homework19;
 
 import java.io.File;
 
+
 public class DirDelete {
     public static void main(String[] args) {
         String path = "/Users/mac/IdeaProjects/BackEndProfM/src/main/java/com/proftelran/homework/homework19";
         try {
             File file = new File(path + "/newdir");
            file.mkdir();
-           deleteDir(file);
         }
         catch (Exception e) {
             e.printStackTrace();
@@ -29,11 +29,8 @@ public class DirDelete {
 
     }
     private static boolean deleteDir (File file) {
-      if (!file.exists()) {
-           return false;
-       }
-        if (file.isDirectory()) {
-            file.delete();
+        if (file.isDirectory() && file.exists()) {
+           file.delete();
         }
         return true;
     }
